fix: agregação diária de métricas usava UTC em vez do fuso do Brasil
ServicoMetricas.IncrementarAsync calculava "hoje" com
DateOnly.FromDateTime(DateTime.UtcNow), então a partir das 21h no
horário do Brasil (UTC-3) o dia já virava para o dia seguinte em UTC,
atribuindo a métrica ao dia errado (ex: domingo 21h já contava como
segunda-feira).

Centraliza o fuso do Brasil em DataHoraHelper (reaproveitando a lógica
que já existia isolada em ProximasMissasService) e usa
DataHoraHelper.HojeBrasil() nos pontos que determinam o "dia de hoje"
para agregação/filtro de métricas.
